move configuration to separate file
